refactor: split portal into pxp library and portal-cli workspace
* Refactors monolithic portal crate into a Cargo workspace with pxp and portal-cli.
* Moves protocol logic (discovery, handshake, manifest, streaming) into pxp library.
