+++ crates/portal-cli/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "hiverra-portal"
+version = "0.12.0"
+authors = ["Spectra010s"]
+edition = "2024"
+repository = "https://github.com/Spectra010s/portal"
+description = "Portal is a lightweight, cross-platform CLI tool built in Rust for sending files and folders between devices instantly."
+homepage = "https://github.com/Spectra010s/portal"
+license = "MIT"
+
+[package.metadata.wix]
+product-name = "Hiverra Portal"
+upgrade-guid = "03150A4A-93A4-4E03-BEF5-5D53BBF4A7C7"
+path-guid = "BD7DEF88-AE13-482C-B366-DA87D1DCAEF1"
+license = true
+license-file = "../../LICENSE"
+eula = false
+
+[[bin]]
+name = "portal"
+path = "src/main.rs"
+
+[dependencies]
+pxp.workspace = true
+clap = { version = "4.5", features = ["derive"] }
+anyhow = "1.0"
+serde = { version = "1.0.228", features = ["derive"] }
+bincode = "1.3.3"
+self_update = { version = "0.42.0", default-features = false, features = ["rustls"] }
+self-replace = "1.5"
+tar = "0.4"
+xz2 = "0.1"
+flate2 = "1.0"
+tokio = { version = "1.49.0", features = ["full"] }
+inquire = "0.9.3"
+reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls-native-roots"] }
+tempfile = "3.25.0"
+toml = "1.0.0"
+home = "0.5.12"
+rand = "0.10.1"
+serde_json = "1.0.149"
+tracing = { version = "0.1.44", features = ["log"] }
+tracing-subscriber = { version = "0.3.23", features = ["fmt", "env-filter"] }
+tracing-appender = "0.2.4"
+indicatif = { version = "0.18.4", features = ["tokio"] }
+chrono = { version = "0.4.38", default-features = false, features = ["clock", "std"] }
