+++ crates/portal-cli/src/config/show.rs
