+++ crates/portal-cli/src/history/filter.rs
