+++ crates/portal-cli/src/history/format.rs
