+++ crates/portal-cli/src/history/json.rs
