+++ crates/portal-cli/src/history/mod.rs
@@ -12,7 +12,7 @@ pub use {
         output_history_json_list,
     },
     models::{
-        HistoryItem, HistoryItemKind, HistoryMode, HistoryStatus, ReceiveSummary,
+        HistoryItem, HistoryItemKind, HistoryMode, HistoryStatus,
         TransferHistoryRecord,
     },
     storage::{append_record, clear_history, delete_history_record, load_history},
