+++ crates/portal-cli/src/update.rs
